\name{ballgownR-package}
\alias{ballgownR-package}
\alias{ballgownR}
\docType{package}
\title{
Manipulation of ballgown output files within R
}
\description{
ballgownR allows you to read in ballgown output files and explore the results. The package might be expanded to deal with derfinder output files as well.
}
\details{
\tabular{ll}{
Package: \tab ballgownR\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2013-04-05\cr
License: \tab What license is it under?\cr
}
Check readGown
}
\author{
Leonardo Collado-Torres, Alyssa Frazee, Andrew Jaffe, Jeffrey Leek

Maintainer: Leonardo Collado-Torres <lcollado@jhsph.edu>
}
\references{
ballgown, derfinder
}
\keyword{ package }
\keyword{ RNAseq }
\keyword{ RNA-seq }
\seealso{
\code{derfinder}
}
\examples{
?readGown
}
